requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
fastapi==0.71.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
imageio==2.13.5
msgpack==1.0.3
neovim==0.3.1
numpy==1.22.1
Pillow==9.0.0
pydantic==1.9.0
pynvim==0.4.3
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
